# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Costs": {
        "properties": {
          "buyer_total_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Buyer Total Paise",
            "type": "integer"
          },
          "freight_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Freight Paise",
            "type": "integer"
          },
          "material_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Material Paise",
            "type": "integer"
          },
          "seller_receives_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Seller Receives Paise",
            "type": "integer"
          }
        },
        "required": [
          "material_paise",
          "freight_paise",
          "buyer_total_paise",
          "seller_receives_paise"
        ],
        "title": "Costs",
        "type": "object"
      },
...
      "Event": {
        "properties": {
          "actor": {
            "enum": [
              "buyer",
              "seller",
              "broker",
              "system"
            ],
            "title": "Actor",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "offer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Id"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "type": {
            "enum": [
              "started",
              "offer",
              "candidate_rejected",
              "agreed",
              "no_deal",
              "failed"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "type",
          "actor",
          "message",
          "created_at"
        ],
        "title": "Event",
        "type": "object"
      },
      "EventsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "title": "Items",
            "type": "array"
          },
          "last_seq": {
            "title": "Last Seq",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "last_seq"
        ],
        "title": "EventsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Negotiation": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_listing_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Listing Id"
          },
          "deal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deal Id"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "max_rounds": {
            "title": "Max Rounds",
            "type": "integer"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/Offer"
            },
            "title": "Offers",
            "type": "array"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "agreed",
              "no_deal",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "requirement_id",
          "status",
          "round",
          "max_rounds",
          "offers",
          "created_at",
          "updated_at"
        ],
        "title": "Negotiation",
        "type": "object"
      },
      "NegotiationAccepted": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "agreed",
              "no_deal",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "title": "NegotiationAccepted",
        "type": "object"
      },
      "NegotiationCreate": {
        "properties": {
          "listing_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Listing Ids",
            "type": "array"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          }
        },
        "required": [
          "requirement_id",
          "listing_ids"
        ],
        "title": "NegotiationCreate",
        "type": "object"
      },
      "NegotiationSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deal Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "agreed",
              "no_deal",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "requirement_id",
          "status",
          "created_at"
        ],
        "title": "NegotiationSummary",
        "type": "object"
      },
      "Offer": {
        "properties": {
          "action": {
            "enum": [
              "propose",
              "counter",
              "accept",
              "reject"
            ],
            "title": "Action",
            "type": "string"
          },
          "author": {
            "enum": [
              "buyer",
              "seller",
              "broker"
            ],
            "title": "Author",
            "type": "string"
          },
          "costs": {
            "$ref": "#/components/schemas/Costs"
          },
          "delivery_at": {
            "format": "date-time",
            "title": "Delivery At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "pickup_at": {
            "format": "date-time",
            "title": "Pickup At",
            "type": "string"
          },
          "quantity_kg": {
            "description": "kilograms",
            "minimum": 0.0,
            "title": "Quantity Kg",
            "type": "integer"
          },
          "responds_to_offer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responds To Offer Id"
          },
          "transport_option_id": {
            "title": "Transport Option Id",
            "type": "string"
          },
          "unit_price_paise_per_tonne": {
            "description": "INR paise per tonne",
            "minimum": 0.0,
            "title": "Unit Price Paise Per Tonne",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "listing_id",
          "transport_option_id",
          "quantity_kg",
          "unit_price_paise_per_tonne",
          "costs",
          "pickup_at",
          "delivery_at",
          "author",
          "action",
          "explanation",
          "expires_at"
        ],
        "title": "Offer",
        "type": "object"
      },
...
      "Page_NegotiationSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NegotiationSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[NegotiationSummary]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessToken": {
        "description": "AUTH_MODE=dev: paste `dev:<email>` (for example `dev:buyer1@demo.bitnbuild.local`). AUTH_MODE=supabase: paste a Supabase access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "B2B exchange for secondary industrial materials. Prices are entered by participants; freight charges are entered or configured estimates. No market price feed, benchmark, or carrier booking is involved.",
    "title": "Karnataka Industrial Byproduct Exchange",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/negotiations": {
      "get": {
        "operationId": "my_negotiations_api_v1_me_negotiations_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NegotiationSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "My Negotiations",
        "tags": [
          "negotiations"
        ]
      }
    },
...
    "/api/v1/negotiations": {
      "post": {
        "operationId": "start_negotiation_api_v1_negotiations_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegotiationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegotiationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Start Negotiation",
        "tags": [
          "negotiations"
        ]
      }
    },
    "/api/v1/negotiations/{negotiation_id}": {
      "get": {
        "operationId": "get_negotiation_api_v1_negotiations__negotiation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "negotiation_id",
            "required": true,
            "schema": {
              "title": "Negotiation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Negotiation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Get Negotiation",
        "tags": [
          "negotiations"
        ]
      }
    },
    "/api/v1/negotiations/{negotiation_id}/events": {
      "get": {
        "operationId": "get_events_api_v1_negotiations__negotiation_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "negotiation_id",
            "required": true,
            "schema": {
              "title": "Negotiation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Seq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Get Events",
        "tags": [
          "negotiations"
        ]
      }
    },
...
  }
}